requirements.txt
fastapi
uvicorn
langchain-core
langchain-google-genai
langgraph
pydantic
python-dotenv
